+++ config/runtime.exs
@@ -15,16 +15,6 @@ end
 config :lightning, :email_addresses,
   admin: System.get_env("EMAIL_ADMIN", "admin@openfn.org")
 
-port = String.to_integer(System.get_env("PORT", "4000"))
-url_port = String.to_integer(System.get_env("URL_PORT", "443"))
-url_scheme = System.get_env("URL_SCHEME", "https")
-
-listen_address =
-  System.get_env("LIGHTNING_LISTEN_ADDRESS", "127.0.0.1")
-  |> String.split(".")
-  |> Enum.map(&String.to_integer/1)
-  |> List.to_tuple()
-
 config :lightning, :adaptor_service,
   adaptors_path: System.get_env("ADAPTORS_PATH", "./priv/openfn")
 
@@ -44,6 +34,7 @@ config :lightning, Oban,
   dispatch_cooldown: 100,
   queues: [
     scheduler: 1,
+    workflow_failures: 1,
     background: 1,
     runs: System.get_env("GLOBAL_RUNS_CONCURRENCY", "1") |> String.to_integer()
   ]
@@ -92,6 +83,54 @@ if System.get_env("MAILGUN_API_KEY") do
     domain: System.get_env("MAILGUN_DOMAIN")
 end
 
+url_port = String.to_integer(System.get_env("URL_PORT", "443"))
+url_scheme = System.get_env("URL_SCHEME", "https")
+
+# The webserver port will always prefer and environment variable _when_
+# given, otherwise it uses the existing config and lastly defaults to 4000.
+port =
+  (System.get_env("PORT") ||
+     Application.get_env(:lightning, LightningWeb.Endpoint)
+     |> Keyword.get(:http, port: nil)
+     |> Keyword.get(:port) ||
+     4000)
+  |> case do
+    p when is_binary(p) -> String.to_integer(p)
+    p when is_integer(p) -> p
+  end
+
+# Binding to loopback ipv4 address prevents access from other machines.
+# http: [ip: {0, 0, 0, 0}, port: 4000],
+# Set `http.ip` to {127, 0, 0, 1} to block access from other machines.
+# Note that this may interfere with Docker networking.
+# Enable IPv6 and bind on all interfaces.
+# Set it to {0, 0, 0, 0, 0, 0, 0, 1} for local network only access.
+# See the documentation on https://hexdocs.pm/plug_cowboy/Plug.Cowboy.html
+# for details about using IPv6 vs IPv4 and loopback vs public addresses.
+listen_address =
+  (System.get_env("LISTEN_ADDRESS") ||
+     Application.get_env(:lightning, LightningWeb.Endpoint)
+     |> Keyword.get(:http, ip: nil)
+     |> Keyword.get(:ip) ||
+     {127, 0, 0, 1})
+  |> case do
+    p when is_binary(p) ->
+      p
+      |> String.split(".")
+      |> Enum.map(&String.to_integer/1)
+      |> List.to_tuple()
+
+    p when is_tuple(p) ->
+      p
+  end
+
+config :lightning, LightningWeb.Endpoint,
+  http: [
+    ip: listen_address,
+    port: port,
+    compress: true
+  ]
+
 if config_env() == :prod do
   database_url =
     System.get_env("DATABASE_URL") ||
@@ -133,16 +172,6 @@ if config_env() == :prod do
 
   config :lightning, LightningWeb.Endpoint,
     url: [host: host, port: url_port, scheme: url_scheme],
-    http: [
-      # Enable IPv6 and bind on all interfaces.
-      # Set it to  {0, 0, 0, 0, 0, 0, 0, 1} for local network only access.
-      # See the documentation on https://hexdocs.pm/plug_cowboy/Plug.Cowboy.html
-      # for details about using IPv6 vs IPv4 and loopback vs public addresses.
-      # ip: {0, 0, 0, 0, 0, 0, 0, 0},
-      # ip: listen_address,
-      port: port,
-      compress: true
-    ],
     secret_key_base: secret_key_base,
     check_origin: origins,
     server: true