+++ lib/lightning/credentials.ex
@@ -26,6 +26,7 @@ defmodule Lightning.Credentials do
 
   def list_credentials(%Project{} = project) do
     Ecto.assoc(project, :credentials)
+    |> preload(:user)
     |> Repo.all()
   end