+++ lib/lightning/projects/project.ex
@@ -18,6 +18,7 @@ defmodule Lightning.Projects.Project do
   @foreign_key_type :binary_id
   schema "projects" do
     field :name, :string
+    field :description, :string
     has_many :project_users, ProjectUser
     has_many :users, through: [:project_users, :user]
     has_many :project_credentials, ProjectCredential
@@ -32,9 +33,10 @@ defmodule Lightning.Projects.Project do
   @doc false
   def changeset(project, attrs) do
     project
-    |> cast(attrs, [:name])
+    |> cast(attrs, [:name, :description])
     |> cast_assoc(:project_users)
     |> validate_required([:name])
+    |> validate_length(:description, max: 240)
     |> validate_format(:name, ~r/^[a-z\-\d]+$/)
   end
 end