+++ lib/lightning_web/live/components/icon.ex
@@ -13,7 +13,11 @@ defmodule LightningWeb.Components.Icon do
 
   def runs(assigns), do: Heroicons.rectangle_stack(assigns)
 
+  def settings(assigns), do: Heroicons.cog_8_tooth(assigns)
+
   def dataclips(assigns), do: Heroicons.cube(assigns)
+
+  def info(assigns), do: Heroicons.information_circle(assigns)
 
   def left(assigns) do
     ~H"""